+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Email notifications for helper tasks:
+  - When someone signs up
+  - When a task is marked as done
+  - When a task is validated
+
+### Changed
+
+- Hierarchical `config.json` for better readability
+
 ## [1.1.0] - 2025-03-16
 
 ### Added